Moved whats new logic to separate file to reduce clutter on extension.ts
